Add strings to the lexer.
This commit extends the lexer to recognize and tokenize string
literals.  It uses the formal Scheme grammar:

⟨string⟩ -> " ⟨string element⟩* "
⟨string element⟩ -> ⟨any character other than " or \⟩ | \" | \\

One problem is that the current REPL implementation assumes that
pressing the "Return" key completes an expression; however, this
is not always true (eg, multi-line proc definitions).  Consequently
pressing "Return" while in the middle of a string literal causes
the interpreter to loop indefinitely.  We should revisit the REPL
Read() mechanism to allow it to handle multi-line expressions.